Made fragment creation into a proper factory
